# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ImmoPro API",
        "description": "Documentation interactive de l'API ImmoPro — plateforme immobilière.\n\n## Authentification\nLa majorité des endpoints nécessitent un token Bearer (Laravel Sanctum).\nConnectez-vous via **POST /api/login** (Admin/Agent) ou **POST /api/client/login** (Client),\npuis cliquez sur le bouton **Authorize** et collez votre token.\n\n## Rôles\n- **admin** : accès complet à l'administration\n- **agent** : gestion des biens assignés et des visites\n- **client** : propriétaires / acheteurs / locataires",
        "contact": {
            "name": "Support ImmoPro",
            "email": "support@immopro.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur principal de l'API ImmoPro"
        }
    ],
    "paths": {
...
        "/resend-otp": {
            "post": {
                "tags": [
                    "Authentification Client"
                ],
                "summary": "Renvoyer un code OTP",
                "description": "Génère et renvoie un nouveau code OTP par email.",
                "operationId": "resendOtp",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "marie.kone@email.com"
                                    },
                                    "pending_token": {
                                        "type": "string",
                                        "example": "xK3mZ9pQ2rL7nY8sV1jW4tU6bO0cE5fA",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nouveau OTP envoyé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Un nouveau code OTP a été envoyé."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucune inscription en attente"
                    }
                }
            }
        },
...
        "/register": {
            "post": {
                "tags": [
                    "Authentification Client"
                ],
                "summary": "Inscription client (étape 1/2)",
                "description": "Valide les données du formulaire, les stocke temporairement en cache et envoie un code OTP à l'email fourni.",
                "operationId": "clientRegister",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "telephone",
                                    "country",
                                    "city",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Marie",
                                        "maxLength": 100
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Koné",
                                        "maxLength": 100
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "marie.kone@email.com"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "+22507654321",
                                        "maxLength": 20
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Côte d'Ivoire",
                                        "maxLength": 100
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Abidjan",
                                        "maxLength": 100
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "MonMotDePasse123",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "MonMotDePasse123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OTP envoyé avec succès — en attente de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Code OTP envoyé à votre email."
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "marie.kone@email.com"
                                        },
                                        "pending_token": {
                                            "type": "string",
                                            "example": "xK3mZ9pQ2rL7nY8sV1jW4tU6bO0cE5fA"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/schemas/ErrorValidation"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorValidation": {
                "title": "Erreur de validation",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "Le champ email est obligatoire."
                            ]
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token Sanctum obtenu après connexion. Format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}